move region code to regions.py, update all tests to reflect; add core gpd function to Region
