+++ src/extremeweatherbench/evaluate.py
@@ -9,7 +9,7 @@ import dacite
 import pandas as pd
 import xarray as xr
 
-from extremeweatherbench import case, config, data_loader, events, utils
+from extremeweatherbench import case, config, data_loader, events, regions, utils
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
@@ -287,13 +287,13 @@ def evaluate(
                 data_class=event,
                 data=yaml_event_case,
                 config=dacite.Config(
-                    type_hooks={utils.Region: utils.map_to_create_region},
+                    type_hooks={regions.Region: regions.map_to_create_region},
                 ),
             )
             # this loop is a workaround to get the location attribute of the case dataclass
             # to be a Region object instead of a dict
             for iter in range(len(cases.cases)):
-                cases.cases[iter].location = utils.map_to_create_region(
+                cases.cases[iter].location = regions.map_to_create_region(
                     cases.cases[iter].location
                 )
         elif isinstance(event, events.EventContainer):
@@ -456,7 +456,9 @@ def get_case_metadata(eval_config: config.Config) -> list[events.EventContainer]
         cases = dacite.from_dict(
             data_class=event,
             data=yaml_event_case,
-            config=dacite.Config(type_hooks={utils.Region: utils.map_to_create_region}),
+            config=dacite.Config(
+                type_hooks={regions.Region: regions.map_to_create_region}
+            ),
         )
         case_metadata_output.append(cases)
     return case_metadata_output
