+++ tests/test_case.py
@@ -2,7 +2,8 @@ import datetime
 
 import pytest
 
-from extremeweatherbench import case, utils
+from extremeweatherbench import case
+from extremeweatherbench.regions import create_region
 
 
 class TestGoodCases:
@@ -12,9 +13,7 @@ class TestGoodCases:
             title="Test Heatwave",
             start_date=datetime.datetime(2000, 1, 1),
             end_date=datetime.datetime(2000, 1, 14),
-            location=utils.create_region(
-                latitude=40, longitude=-100, bounding_box_degrees=5
-            ),
+            location=create_region(latitude=40, longitude=-100, bounding_box_degrees=5),
             event_type="heat_wave",
         )
         subset_dataset = heatwave_case.perform_subsetting_procedure(
@@ -29,9 +28,7 @@ class TestGoodCases:
             title="Test Case",
             start_date=datetime.datetime(2000, 1, 1),
             end_date=datetime.datetime(2000, 1, 14),
-            location=utils.create_region(
-                latitude=40, longitude=-100, bounding_box_degrees=5
-            ),
+            location=create_region(latitude=40, longitude=-100, bounding_box_degrees=5),
             event_type="heat_wave",
         )
         valid_case = {
@@ -39,7 +36,7 @@ class TestGoodCases:
             "title": "Test Case",
             "start_date": datetime.datetime(2000, 1, 1),
             "end_date": datetime.datetime(2000, 1, 14),
-            "location": utils.create_region(
+            "location": create_region(
                 latitude=40, longitude=-100, bounding_box_degrees=5
             ),
             "event_type": "heat_wave",
