+++ tests/test_evaluate.py
@@ -5,7 +5,8 @@ import pandas as pd
 import pytest
 import xarray as xr
 
-from extremeweatherbench import case, evaluate, events, utils
+from extremeweatherbench import case, evaluate, events
+from extremeweatherbench.regions import create_region
 
 
 def test_get_case_metadata(sample_config):
@@ -20,9 +21,7 @@ def test_evaluate_individualcase(sample_forecast_dataset, sample_gridded_obs_dat
         title="test_case",
         start_date=datetime.datetime(2021, 6, 20),
         end_date=datetime.datetime(2021, 7, 3),
-        location=utils.create_region(
-            latitude=45, longitude=-100, bounding_box_degrees=5
-        ),
+        location=create_region(latitude=45, longitude=-100, bounding_box_degrees=5),
         event_type="heat_wave",
         data_vars=["2m_temperature"],
     )
