+++ tests/test_utils.py
@@ -6,6 +6,10 @@ import pytest
 import xarray as xr
 
 from extremeweatherbench import utils
+from extremeweatherbench.regions import (
+    clip_dataset_to_bounding_box_degrees,
+    create_region,
+)
 
 
 @pytest.mark.parametrize(
@@ -168,30 +172,28 @@ def test_clip_dataset_to_bounding_box_degrees():
     )
 
     # Test case 1: Single value for box_degrees, latitude ascending
-    location_center = utils.create_region(
-        latitude=40, longitude=100, bounding_box_degrees=10
-    )
-    clipped_ds = utils.clip_dataset_to_bounding_box_degrees(ds, location_center)
+    location_center = create_region(latitude=40, longitude=100, bounding_box_degrees=10)
+    clipped_ds = clip_dataset_to_bounding_box_degrees(ds, location_center)
     assert np.all(clipped_ds.latitude >= 35)
     assert np.all(clipped_ds.latitude <= 45)
     assert np.all(clipped_ds.longitude >= 95)
     assert np.all(clipped_ds.longitude <= 105)
 
     # Test case 2: Tuple for box_degrees, latitude ascending
-    location_center = utils.create_region(
+    location_center = create_region(
         latitude=40, longitude=100, bounding_box_degrees=(5, 10)
     )
-    clipped_ds = utils.clip_dataset_to_bounding_box_degrees(ds, location_center)
+    clipped_ds = clip_dataset_to_bounding_box_degrees(ds, location_center)
     assert np.all(clipped_ds.latitude >= 37.5)
     assert np.all(clipped_ds.latitude <= 42.5)
     assert np.all(clipped_ds.longitude >= 95)
     assert np.all(clipped_ds.longitude <= 105)
 
     # Test case 3: Negative longitude, latitude ascending
-    location_center = utils.create_region(
+    location_center = create_region(
         latitude=40, longitude=-100, bounding_box_degrees=10
     )
-    clipped_ds = utils.clip_dataset_to_bounding_box_degrees(ds, location_center)
+    clipped_ds = clip_dataset_to_bounding_box_degrees(ds, location_center)
     assert np.all(clipped_ds.latitude >= 35)
     assert np.all(clipped_ds.latitude <= 45)
     assert np.all(clipped_ds.longitude >= 255)  # -100 + 360 - 5 = 255
@@ -199,34 +201,28 @@ def test_clip_dataset_to_bounding_box_degrees():
 
     # Test case 4: Latitude descending
     ds_desc = ds.reindex(latitude=ds.latitude[::-1])
-    location_center = utils.create_region(
-        latitude=40, longitude=100, bounding_box_degrees=10
-    )
-    clipped_ds = utils.clip_dataset_to_bounding_box_degrees(ds_desc, location_center)
+    location_center = create_region(latitude=40, longitude=100, bounding_box_degrees=10)
+    clipped_ds = clip_dataset_to_bounding_box_degrees(ds_desc, location_center)
     assert np.all(clipped_ds.latitude >= 35)
     assert np.all(clipped_ds.latitude <= 45)
     assert np.all(clipped_ds.longitude >= 95)
     assert np.all(clipped_ds.longitude <= 105)
 
     # Test wrapping around prime meridian
-    location_center = utils.create_region(
-        latitude=0, longitude=0, bounding_box_degrees=10
-    )
-    clipped_ds = utils.clip_dataset_to_bounding_box_degrees(ds, location_center)
+    location_center = create_region(latitude=0, longitude=0, bounding_box_degrees=10)
+    clipped_ds = clip_dataset_to_bounding_box_degrees(ds, location_center)
 
     assert np.any(clipped_ds.longitude < 10)
     assert np.any(clipped_ds.longitude > 350)
 
     # Test case 5: Edge cases, ensuring no errors when clipping at boundaries
-    location_center = utils.create_region(
-        latitude=90, longitude=100, bounding_box_degrees=10
-    )
-    clipped_ds = utils.clip_dataset_to_bounding_box_degrees(ds, location_center)
+    location_center = create_region(latitude=90, longitude=100, bounding_box_degrees=10)
+    clipped_ds = clip_dataset_to_bounding_box_degrees(ds, location_center)
 
-    location_center = utils.create_region(
+    location_center = create_region(
         latitude=-90, longitude=100, bounding_box_degrees=10
     )
-    clipped_ds = utils.clip_dataset_to_bounding_box_degrees(ds, location_center)
+    clipped_ds = clip_dataset_to_bounding_box_degrees(ds, location_center)
 
 
 def test_align_point_obs_from_gridded_basic(
